pyproject.toml
[tool.poetry]
name = "chmsapi"
version = "0.1.0"
description = ""
authors = ["Festus Kwafo Ofori <68119922+Festus-Kwafo@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "api", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
mysqlclient = "^2.2.3"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
path = "^16.10.0"
json-log-formatter = "^0.5.2"
asyncmy = "^0.2.9"
python-ulid = "^2.2.0"
python-multipart = "^0.0.9"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
fastapi-pagination = "^0.12.15"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
email-validator = "^2.1.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
